confirm_otg_storage_access_text+1
I have not translated 'OTG' though I do not believe that it is a common known acronym, at least not among Danes in general. I do not know it myself. Acronymfinder.com suggests: On-The-Go (USB 2.0 specification), On the Ground (sales) and Off the Ground (gaming) - are you talking about USB2? Maybe I could have omitted the term 'OTG' and just translate 'device' - or will it make sense for those who need to grant access to the root directory?