Update nl string.xml
Made more in style of official spelling of Dutch language. (days of the week are written with small letter in Dutch, shorten form is with 2 letters). The letter "ij" in vrijdag is one symbol and cannot be splitted.

This should make the clock widget more in style of the "official" android clock widget.