Minor fixes in Spanish strings
This assumes that the `saving_label` is always used as a header section for saving related settings rather than some indication that something is actually saving. By the way, if this is really the case, I think many of the translations are making similar conjugation errors with these settings header labels.