Truncate all output files
Address the behavior change from Android bug 180526528 so that
outputting to a URI that already exists replaces the entire destination
instead of only modifying part of it, when the output is smaller.

This resolves an issue with Simple Gallery in which resizing a file
to a smaller size when that app only has Media Management permission
(as opposed to All Files permission) leaves extra data at the end,
with the file not getting any smaller.